+++ HISTORY.md
@@ -2,6 +2,9 @@
 
 ## in development
 
+###2018-12-28
+* Blocks, Threads: added 'neg' selector to monad operator dropdown
+
 ### 2018-12-07
 * GUI: fixed #2284 (executing "inherit" should not un-hide palette in presentation mode)