+++ HISTORY.md
@@ -5,6 +5,7 @@
 ### 2019-01-02
 * Objects, Blocks, Threads, GUI, Store: added third color dimension and transparency to pen
 * renamed help screens for "setHue" to "setPenHSVA" and "changeHue" to "changePenHSVA"
+* pushed dev version to 4.2.3
 
 ### 2019-01-01
 * Morphic: added HSL color conversion support